+++ packages/better-auth/src/plugins/organization/adapter.ts
@@ -1146,10 +1146,11 @@ export const getOrgAdapter = <O extends OrganizationOptions>(
 			if (data.fromStatus) {
 				where.push({ field: "status", value: data.fromStatus });
 			}
-			const invitation = await adapter.update<InferInvitation<O, false>>({
+			const invitation = await adapter.incrementOne<InferInvitation<O, false>>({
 				model: "invitation",
 				where,
-				update: {
+				increment: {},
+				set: {
 					status: data.status,
 				},
 			});